+++ crates/aptos/CHANGELOG.md
@@ -2,10 +2,14 @@
 
 All notable changes to the Aptos CLI will be captured in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and the format set out by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
-### Updated
+## [2.0.3] - 2023/08/02
+### Fixed
+- Fixed the following input arguments issue when running `aptos move view`
+  - #8513: Fixed issue where CLI does not work with big numbers
+  - #8982: Fixed args issue when passing in u64/u128/u256 parameters
+### Update
+- CLI documentation refactor
 - Updated CLI source compilation to use rust toolchain version 1.71.0 (from 1.70.0).
-
 ### Fixed
 * Verify package now does not fail on a mismatched upgrade number